Moving all of the mock declarations to their own file so the tests are less cluttered.
